@@ -1,5 +1,9 @@
 import { addRecipe } from "@/schemas/recipe";
-import { CookingMethod, MealType, Nationality } from "@prisma/client";
+import {
+  type CookingMethod,
+  type MealType,
+  type Nationality,
+} from "@prisma/client";
 
 export type DropdownListValues = MealType | CookingMethod | Nationality;
 export type ListInputFields = keyof Pick<addRecipe, "ingredients" | "steps">;
