@@ -1,6 +1,6 @@
 // src/server/db/client.ts
 import { PrismaClient } from "@prisma/client";
-import { env } from "../env.mjs";
+import { env } from "~/env.js";
 
 declare global {
   var prisma: PrismaClient | undefined;
